+++ LyndaCoursesDownloader.DownloaderConfig/LyndaCoursesDownloader.DownloaderConfig.csproj
@@ -1,0 +1,15 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>netstandard2.0</TargetFramework>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Newtonsoft.Json" Version="12.0.3" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\LyndaCoursesDownloader.CourseContent\LyndaCoursesDownloader.CourseContent.csproj" />
+  </ItemGroup>
+
+</Project>
